Use a discriminant union
